requirements.txt
streamlit
pandas
numpy
tensorflow
matplotlib
Pillow
mysql.connector
requests
cv2
plotly
schedule
opencv-python


